reduce occurrence of rebreaking
There's a bug where the initial breakpoint can be hit even after it's
been removed on a highly threaded target, around 1/50 chance. I believe
this is due to the target being moved between cores and there being no
syncronisation on the memory changes when we write the breakpoint
instruction. The suspension and resumption of the the task reduces the
frequency of, but does not eliminate, this bug. So in other words this
commit doesn't do much and I have plans to switch to using hardware
breakpoints always.