darwin: fix extra thread attacher getting stuck
It seems that sometimes the extra thread that tries to attach to
the remaining process threads doesn't have it's kevent64 call
interrupted when the thread receives a signal. We now at least listen
for the process itself receiving a signal and end ourselves when we
are that thread.

We also add some the uncommitted manual tests we've been using for
a little while.