linux: use Py_Main instead of libc start to write syscall instruction
On amd64 the build-id of libc.so.6 was loaded into a readonly page not
an executable page. The start of the executable part of the library
contains the .plt section.
The idea of using the start of libc came from kubo/injector which
has to be very general but since we are tracing Python we can use
knowledge of how the python binary is generally used in order to select
a safe spot to overwrite an instruction or two.