linux: remove breakpoints from other threads when executing injection code
When running the injection code, the captive thread may release the GIL
in order to do IO. Another thread can acquire the GIL in this time.
If the breakpoint is enabled in that thread it will stop and wait for
the tracer but the tracer is waiting for the injection code to finish
executing.

We solve the problem by removing breakpoints before and readding them
after executing the injection code.

Another option might have been to pause the threads who have breakpoints
but when testing that method I found the paused thread then was
dying of signal death and I wasn't able to deduce why. Therefore this
seems to be the valid solution.

It needs a bit of refactoring now. That can come in a future commit.