send maps via file to reduce chance of lockup
If the tracee writes too much data during setuntrace, the unix socket
buffer may fill up block the tracee and cause a deadlock.

A contributor to this problem was the sending of maps when aggregation
is done. For now we send maps file an additional file at the end to
avoid blocking up the socket. If the user has written END probes that
do something similar they will still experience the issue for now.