tracee: store each PMTMap as a file
This is a first iteration, starting to feel out the right shape
for an out-of-process aggregation buffer.

Next steps will be to notify the tracer of it at creation time,
and then have the tracer start to read it.

Once something feels like it's working, we'll migrate it into
C and an mmap'd buffer so we can avoid system call overhead.