share aggregation buffers by filename at creation time
The tracee now sends a message with the filename of the aggregation
buffer at the time of creation. This means that if the tracee dies,
the tracer will still have access to aggregation data that was
emitted before the death.

Next step may be to split the buffer and start to clean it when
it gets too large.